+++ webpack/develop.js
@@ -35,7 +35,7 @@ const config = merge(common, {
         hot: true,
         // host: '0.0.0.0',
         // hot 和 hotOnly 的区别是在某些模块不支持热更新的情况下，前者会自动刷新页面，后者不会刷新页面，而是在控制台输出热更新失败
-        hotOnly:true,
+        // hotOnly: true,
         proxy: {
             '/api': {
                 target: 'http://domain.hostname.com',