+++ webpack/production.js
@@ -27,7 +27,7 @@ const config = merge(common, {
     output: {
         path: path.resolve(ROOT_PATH, 'dist'),
         filename: 'js/[name]-[chunkhash:8].js',
-        publicPath: '/',   //  插入到html模板的路径前缀
+        publicPath: './',   //  插入到html模板的路径前缀
     },
     // devtool: 'cheap-module-source-map',
     bail: true, //  在遇到错误的时候，打包过程将会退出，且把错误信息打印到 stderr
@@ -141,7 +141,7 @@ const config = merge(common, {
                 // cache: '.cache', //  缓存路径
                 parallel: true, //  多线程压缩
                 sourceMap: true,    //开启sourceMap
-                extractComments: true,  //  提取出注释
+                // extractComments: true,  //  提取出注释（自测有问题，干掉vendor.js的(window.webpackJsonp=window.webpackJsonp||[])）
                 // 过滤命令行控制台警告输出
                 // warningsFilter: (warning, source) => {
                 //     // console.log('warning ==== >', warning, source)